BackEnd/Marine-Threat-Detection/requirements.txt
Flask==2.3.2
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.6
